Add user authentication

+++ studybud/base/views.py
@@ -1,15 +1,65 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, HttpResponse
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import login, logout, authenticate
+from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib import messages
+
 from .models import Room, Topic
 from .forms import RoomForm
 
+
 # Create your views here.
 
-rooms = [
-    {"id": 1, "name": "Let's learn python"},
-    {"id": 2, "name": "Design with me"},
-    {"id": 3, "name": "Frontend developers"},
-]
+
+def login_page(request):
+    page = "login"
+
+    if request.user.is_authenticated:
+        return redirect("home")
+
+    if request.method == "POST":
+        username = request.POST.get("username").lower()
+        password = request.POST.get("password")
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user=user)
+            return redirect("home")
+        else:
+            messages.error(request, "Username or password is incorrect")
+
+    return render(request, "base/login_register.html", {"page": page})
+
+
+def logout_user(request):
+    logout(request)
+    return redirect("home")
+
+
+def register_page(request):
+    page = "register"
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "An error occurred during registration")
+
+    return render(request, "base/login_register.html", {"page": page, "form": form})
 
 
 def home(request):
@@ -36,6 +86,7 @@ def room(request, pk):
     return render(request, "base/room.html", context=context)
 
 
+@login_required(login_url="/login")
 def create_room(request):
     form = RoomForm()
 
@@ -48,9 +99,13 @@ def create_room(request):
     return render(request, "base/room_form.html", context={"form": form})
 
 
+@login_required(login_url="/login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to update this room.")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -61,8 +116,13 @@ def update_room(request, pk):
     return render(request, "base/room_form.html", context={"form": form})
 
 
+@login_required(login_url="/login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to delete this room.")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
